+++ examples/scripts/ct_astra_noreg_pcg.py
@@ -30,7 +30,7 @@ from xdesign import Foam, discrete_phantom
 
 from scico import loss, plot
 from scico.linop import CircularConvolve
-from scico.linop.radon_astra import XRayTransform
+from scico.linop.xray_astra import XRayTransform
 from scico.solver import cg
 
 """
