+++ examples/scripts/ct_astra_odp_train_foam2.py
@@ -58,7 +58,7 @@ from scico import flax as sflax
 from scico import metric, plot
 from scico.flax.examples import load_ct_data
 from scico.flax.train.traversals import clip_positive, construct_traversal
-from scico.linop.radon_astra import XRayTransform
+from scico.linop.xray_astra import XRayTransform
 
 """
 Prepare parallel processing. Set an arbitrary processor count (only
