+++ examples/scripts/ct_astra_weighted_tv_admm.py
@@ -29,7 +29,7 @@ from xdesign import Soil, discrete_phantom
 
 import scico.numpy as snp
 from scico import functional, linop, loss, metric, plot
-from scico.linop.radon_astra import XRayTransform
+from scico.linop.xray_astra import XRayTransform
 from scico.optimize.admm import ADMM, LinearSubproblemSolver
 from scico.util import device_info
