+++ examples/scripts/ct_fan_svmbir_ppp_bm3d_admm_prox.py
@@ -37,7 +37,7 @@ import scico.numpy as snp
 from scico import metric, plot
 from scico.functional import BM3D
 from scico.linop import Diagonal, Identity
-from scico.linop.radon_svmbir import SVMBIRExtendedLoss, XRayTransform
+from scico.linop.xray_svmbir import SVMBIRExtendedLoss, XRayTransform
 from scico.optimize.admm import ADMM, LinearSubproblemSolver
 from scico.util import device_info
