+++ examples/scripts/ct_svmbir_tv_multi.py
@@ -31,7 +31,7 @@ from xdesign import Foam, discrete_phantom
 import scico.numpy as snp
 from scico import functional, linop, metric, plot
 from scico.linop import Diagonal
-from scico.linop.radon_svmbir import SVMBIRSquaredL2Loss, XRayTransform
+from scico.linop.xray_svmbir import SVMBIRSquaredL2Loss, XRayTransform
 from scico.optimize import PDHG, LinearizedADMM
 from scico.optimize.admm import ADMM, LinearSubproblemSolver
 from scico.util import device_info
