+++ scico/linop/xray_svmbir.py
